Implement get all permissions endpoint

+++ security-server/security-server-dto/src/main/java/by/kovzov/uis/security/dto/RoleDto.java
@@ -22,5 +22,5 @@ public class RoleDto {
     String name;
 
     @NotNull
-    List<Integer> permissionIds;
+    List<Long> permissionIds;
 }
